+++ appengine/flexible/django_cloudsql/mysite/settings.py
@@ -116,7 +116,7 @@ else:
     DATABASES['default']['HOST'] = '127.0.0.1'
 # [END dbconfig]
 
-# Use a in-memort sqlite3 database when testing in Kokoro
+# Use a in-memory sqlite3 database when testing in CI systems
 if os.getenv('TRAMPOLINE_CI', None):
     DATABASES = {
         'default': {
