--- appengine/flexible/django_cloudsql/pytest.ini
@@ -1,3 +1,0 @@
-[pytest]
-DJANGO_SETTINGS_MODULE = mysite.settings
-python_files = tests.py
